Fix external tests
