+++ crates/agent2/Cargo.toml
@@ -10,7 +10,7 @@ path = "src/agent2.rs"
 
 [features]
 test-support = ["db/test-support"]
-e2e = ["agent_servers/e2e"]
+e2e = []
 
 [lints]
 workspace = true
@@ -73,7 +73,7 @@ zstd.workspace = true
 
 [dev-dependencies]
 agent = { workspace = true, "features" = ["test-support"] }
-agent_servers = { workspace = true, "features" = ["e2e", "test-support"] }
+agent_servers = { workspace = true, "features" = ["test-support"] }
 assistant_context = { workspace = true, "features" = ["test-support"] }
 ctor.workspace = true
 client = { workspace = true, "features" = ["test-support"] }
