+++ crates/agent_servers/Cargo.toml
@@ -6,8 +6,8 @@ publish.workspace = true
 license = "GPL-3.0-or-later"
 
 [features]
-test-support = ["acp_thread/test-support", "gpui/test-support", "project/test-support"]
-e2e = ["dep:env_logger", "fs"]
+test-support = ["acp_thread/test-support", "gpui/test-support", "project/test-support", "dep:env_logger", "fs", "client/test-support", "dep:gpui_tokio", "reqwest_client/test-support"]
+e2e = []
 
 [lints]
 workspace = true
@@ -23,12 +23,14 @@ agent-client-protocol.workspace = true
 agent_settings.workspace = true
 agentic-coding-protocol.workspace = true
 anyhow.workspace = true
+client = { workspace = true, optional = true }
 collections.workspace = true
 context_server.workspace = true
 env_logger = { workspace = true, optional = true }
 fs = { workspace = true, optional = true }
 futures.workspace = true
 gpui.workspace = true
+gpui_tokio = { workspace = true, optional = true }
 indoc.workspace = true
 itertools.workspace = true
 language.workspace = true
@@ -38,6 +40,7 @@ log.workspace = true
 paths.workspace = true
 project.workspace = true
 rand.workspace = true
+reqwest_client = { workspace = true, optional = true }
 schemars.workspace = true
 semver.workspace = true
 serde.workspace = true
@@ -59,9 +62,12 @@ libc.workspace = true
 nix.workspace = true
 
 [dev-dependencies]
+client = { workspace = true, features = ["test-support"] }
 env_logger.workspace = true
 fs.workspace = true
 language.workspace = true
 indoc.workspace = true
 acp_thread = { workspace = true, features = ["test-support"] }
 gpui = { workspace = true, features = ["test-support"] }
+gpui_tokio.workspace = true
+reqwest_client = { workspace = true, features = ["test-support"] }
